// src/app/Model/Contrato.ts
export class Contrato {
    idContrato: Number = 0;     
    titulo: String = ""         
    descicao: String  = "";     
    codigo: String  = "";         
    nome: String = "";            
    dtFimVigencia: String = "";  
    dtInicioVigencia: String = ""
    modalidade: String   = "";   
    url: String  = "";     
    status: Number = 0;     
}